+++ src/interfaces/plugins/index.ts
@@ -1,5 +1,6 @@
 export * from './interactions/interactions.interface';
 export * from './keyboardHandler/keyboardHandler.interface';
+export * from './keyboardHandler/simpleKeyboardHandler.interface';
 export * from './liveSearch/liveSearch.interface';
 export * from './normalState/normalState.interface';
 export * from './normalState/simpleNormalState.interface';