+++ src/interfaces/plugins/keyboardHandler/simpleKeyboardHandler.interface.ts
@@ -1,0 +1,14 @@
+import {KeyboardHandlerOptions} from './keyboardHandler.interface';
+
+/**
+ * Options for simple keyboard handler plugin
+ */
+export interface SimpleKeyboardHandlerOptions extends KeyboardHandlerOptions
+{
+    //######################### properties #########################
+
+    /**
+     * Debounce timeout for selecting first option when typing
+     */
+    selectFirstDebounceTimeout: number;
+}