+++ src/plugins/liveSearch/noLiveSearch/noLiveSearch.component.ts
@@ -25,20 +25,4 @@ export class NoLiveSearchComponent implements LiveSearch
     {
 
     }
-
-    //######################### public methods - implementation of SelectPlugin #########################
-
-    /**
-     * @inheritdoc
-     */
-    public initialize(): void
-    {
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public initOptions(): void
-    {
-    }
 }