+++ src/plugins/valueHandler/staticValueHandler/staticValueHandler.component.ts
@@ -94,22 +94,6 @@ export class StaticValueHandler<TValue = unknown> implements ValueHandler<TValue
         });
     }
 
-    //######################### public methods - implementation of SelectPlugin #########################
-
-    /**
-     * @inheritdoc
-     */
-    public initialize(): void
-    {
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public initOptions(): void
-    {
-    }
-
     //######################### public methods - implementation of ValueHandler #########################
 
     /**